requirements.txt
openai
pinecone
python-dotenv
langchain_community
langchain-openai
langchain-huggingface
langchain-google-genai
langchain-pinecone
llama-index
streamlit
fastapi
pypdf

